Clearing lines works
There are several bugs: buttons receive faulty toggles often, and while
this effect has been mitigated, it still poses issues. If a Tetromino
hits another, but the user moves it to a space where it can fall freely
again, it may not always continue to drop. This is likely due to the
behavior of the runnning threads.